requirements.txt
torch==2.3.0+cu121 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.18.0+cu121 --index-url https://download.pytorch.org/whl/cu121
numpy==2.4.1
matplotlib==3.9.0
tqdm==4.66.2
opencv-python==4.13.0.90
scikit-image==0.26.0
mamba_ssm==2.3.0
causal_conv1d==1.6.0
transformers==4.57.6
einops==0.8.1
triton==3.6.0
scipy==1.17.0
pillow==10.3.0
basicsr==1.4.2